auto-create omni root folder, api to download manifest, tests

+++ src/main/java/omnidrive/api/dropbox/DropboxAccount.java
@@ -20,11 +20,37 @@ public class DropboxAccount extends BaseAccount {
 
     @Override
     protected void createRootFolder() throws BaseException {
-        try {
-            this.client.createFolder(ROOT_FOLDER_PATH);
-        } catch (DbxException ex) {
-            throw new DropboxException("Failed to create folder.");
+        if (!isOmniDriveFolderExists()) {
+            try {
+                DbxEntry.Folder rootFolder = this.client.createFolder(OMNIDRIVE_ROOT_FOLDER_PATH);
+                if (rootFolder == null) {
+                    throw new DropboxException("Failed to create folder.");
+                } else if (!rootFolder.isFolder()) {
+                    throw new DropboxException("Failed to create folder.");
+                }
+            } catch (DbxException ex2) {
+                throw new DropboxException("Failed to create folder.");
+            }
         }
+    }
+
+    @Override
+    public String getOmniDriveFolderId() throws BaseException {
+        if (this.omniDriveFolderId != null)
+            return this.omniDriveFolderId;
+
+        try {
+            DbxEntry rootEntry = this.client.getMetadata(OMNIDRIVE_ROOT_FOLDER_PATH);
+            if (rootEntry != null) {
+                if (rootEntry.isFolder()) {
+                    this.omniDriveFolderId = rootEntry.path;
+                }
+            }
+        } catch (DbxException ex1) {
+            this.omniDriveFolderId = null;
+        }
+
+        return this.omniDriveFolderId;
     }
 
     @Override
@@ -59,7 +85,9 @@ public class DropboxAccount extends BaseAccount {
 
         try {
             DbxEntry.File file = this.client.uploadFile(getFullPath(name), DbxWriteMode.add(), size, inputStream);
-            fileName = file.asFile().name;
+            if (file != null) {
+                fileName = file.asFile().name;
+            }
         } catch (FileNotFoundException ex) {
             throw new DropboxException("Input file not found.");
         } catch (DbxException ex) {
@@ -105,10 +133,14 @@ public class DropboxAccount extends BaseAccount {
     public void updateFile(String name, InputStream inputStream, long size) throws BaseException {
         try {
             DbxEntry entry = this.client.getMetadata(getFullPath(name));
-            if (entry.isFile()) {
-                this.client.uploadFile(entry.asFile().path, DbxWriteMode.update(entry.asFile().rev), size, inputStream);
+            if (entry != null) {
+                if (entry.isFile()) {
+                    this.client.uploadFile(entry.asFile().path, DbxWriteMode.update(entry.asFile().rev), size, inputStream);
+                } else {
+                    throw new DropboxException("Not a file.");
+                }
             } else {
-                throw new DropboxException("Not a file.");
+                throw new DropboxException("File does not exist.");
             }
         } catch (IOException ex) {
             throw new DropboxException("Failed to update file: " + ex.getMessage());
@@ -121,14 +153,40 @@ public class DropboxAccount extends BaseAccount {
     public void removeFolder(String name) throws BaseException {
         try {
             DbxEntry entry = this.client.getMetadata(getFullPath(name));
-            if (entry.isFolder()) {
-                this.client.delete(getFullPath(name));
+            if (entry != null) {
+                if (entry.isFolder()) {
+                    this.client.delete(getFullPath(name));
+                } else {
+                    throw new DropboxException("Not a folder.");
+                }
             } else {
-                throw new DropboxException("Not a folder.");
+                throw new DropboxException("Folder ndoes not exist.");
             }
         } catch (DbxException ex) {
             throw new DropboxException(ex.getMessage());
         }
+    }
+
+    @Override
+    public long downloadManifestFile(OutputStream outputStream) throws BaseException {
+        long size = 0;
+
+        if (!isOmniDriveFolderExists()) {
+            throw new DropboxException("No 'OmniDrive' root folder exists");
+        }
+
+        try {
+            DbxEntry.File dbxFile = this.client.getFile(getFullPath("manifest"), null, outputStream);
+            if (dbxFile != null) {
+                size = dbxFile.numBytes;
+            }
+        } catch (DbxException ex) {
+            throw new DropboxException("Failed to download manifest file.");
+        } catch (IOException ex) {
+            throw new DropboxException("Failed to find manifest file.");
+        }
+
+        return size;
     }
 
     @Override
